Check id existence before delete

+++ src/main/java/com/imb2025/calificaciones/service/jpa/DocenteServiceImpl.java
@@ -63,10 +63,11 @@ public class DocenteServiceImpl implements IDocenteService {
     }
 
     @Override
-    public void deleteById(Long id) {
-
-            repo.deleteById(id);
-
+    public void deleteById(Long id) throws Exception {
+        if (!repo.existsById(id)) {
+            throw new Exception("No se puede eliminar el id: " + id + " porque no existe");
+        }
+        repo.deleteById(id);
     }
 
 }
