+++ src/main/java/com/imb2025/calificaciones/service/jpa/EstadoCursadaServiceImpl.java
@@ -43,7 +43,10 @@ public class EstadoCursadaServiceImpl implements IEstadoCursadaService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws Exception {
+        if (!repository.existsById(id)) {
+            throw new Exception("No se puede eliminar el id: " + id + " porque no existe");
+        }
         repository.deleteById(id);
     }
